refactor: inode allocation
Motivation: In lab1, we have only one type (language-level) of inode.
But in lab2, the metadata server has another type of inode: regular inode.

To make the code clean and reusable, templates and functional programming are used.

Achievement: interfaces for lab2 are independent with those for lab1.
To denote similar methods for different labs, I append suffix "metadata_server"
if the interface should be called by the class `MetadataServer`.