fix: InodeManager::get_type
The class `InodeManager` used to get the type of an inode by directly
"reinterpret" the byte data to `Inode*`, which is tricky and efficient.

However, since we need to extend the inode type (e.g. RegularInode for lab2),
the method should be abandoned because its lack of scalability.

Now we obtain the inode type by read the first 4 bytes (`sizeof(InodeType)`).
As long as the storage layout of inode stays the same, we can always get
the correct inode type.